Twig(link)Amended the link between homepage and trick page and added display for category+author
