+++ templates/community/home.html.twig
@@ -16,7 +16,7 @@
                 <a href="/trick/{{ trick.id }}">Lire la suite ...</a>
                 </div>
                 <img src="{{ trick.imageLink }}" alt="{{ trick.name }}"/>
-                <p class="list-group-item">Créé par  || Catégorie :
+                <p class="list-group-item">Créé par {{ trick.user.username }} || Catégorie : {{ trick.category.name }}
                 </p>
             </div>
         {% endfor %}
