+++ templates/community/viewTrick.html.twig
@@ -9,7 +9,7 @@
             <div>{{ trick.description | raw }}
             </div>
             <img src="{{ trick.imageLink }}" alt="{{ trick.name }}"/>
-            <p class="list-group-item">Créé par || Catégorie :
+            <p class="list-group-item">Créé par {{ trick.user.username }} || Catégorie : {{ trick.category.name }}
             </p>
             <p><a href="/">Retour</a></p>
         </div>
